Move npm package code into dedicated folder
We want to get rid of webpacker
