+++ package/src/__tests__/i18n.spec.js
