+++ package/src/i18n.js
