+++ package/src/node_tree.js
